part1/requirements.txt
# 音频处理与采集
pyaudio>=0.2.11
webrtcvad>=2.0.10
numpy>=1.21.0

# 语音转文本
openai-whisper>=20231117
opencc-python-reimplemented>=1.1.6

# 大模型对话
requests>=2.28.0

# 控制台交互
keyboard>=0.13.5
rich>=13.0.0
prompt-toolkit>=3.0.36